requirements.txt
Django==5.1.1
django-ninja==1.3.0
pytest==8.3.2
pytest-django==4.9.0
django-auth==0.1.9
